added script for put requests for questions
